requirements.txt
python-dotenv
watchdog
pandas
openpyxl
sqlalchemy
psycopg2-binary
python-dateutil
pywin32
pyperclip